@@ -10,7 +10,6 @@ try:
 except ImportError:
     pass
 from opentrons.config import pipette_config, robot_configs, feature_flags
-from opentrons import hardware_control
 from opentrons.types import Mount, Point
 from opentrons.hardware_control.types import CriticalPoint
 from . import jog, position, dots_set, z_pos
@@ -85,12 +84,8 @@ class SessionManager:
         self.z_value = None
         self.cp = None
         self.pipette_id = None
-        self.adapter = hardware
+        self.adapter = hardware.sync
         self.current_transform = identity_deck_transform()
-
-        if feature_flags.use_protocol_api_v2():
-            self.adapter = hardware_control.adapters.SynchronousAdapter(
-                hardware)
 
         robot_configs.backup_configuration(self.adapter.config)
         # Start from fresh identity matrix every calibration session
