@@ -5,7 +5,8 @@ import logging
 from typing import Any, Callable, Optional, TYPE_CHECKING, Union
 
 from opentrons.protocols.types import APIVersion
-from opentrons.hardware_control import types, adapters, API, HardwareAPILike
+from opentrons.hardware_control import (types, adapters, API,
+                                        HardwareAPILike, ThreadManager)
 
 if TYPE_CHECKING:
     from .contexts import InstrumentContext
@@ -139,12 +140,8 @@ class AxisMaxSpeeds(UserDict):
     """
 
     def __getitem__(self, key: Union[str, types.Axis]):
-        if key in types.Axis:
-            return self.data[key]
-        elif isinstance(key, str):
-            return self.data[types.Axis[key.upper()]]
-        else:
-            raise KeyError(key)
+        checked_key = AxisMaxSpeeds._verify_key(key)
+        return self.data[checked_key]
 
     @staticmethod
     def _verify_key(key: Any) -> types.Axis:
@@ -186,60 +183,35 @@ class AxisMaxSpeeds(UserDict):
 
 class HardwareManager:
     def __init__(self, hardware):
-        if None is hardware:
-            self._is_orig = True
-            self._built_own_adapter = True
-            self._current = adapters.SynchronousAdapter.build(
-                API.build_hardware_simulator)
+        if hardware is None:
+            self._current = ThreadManager(API.build_hardware_simulator).sync
         elif isinstance(hardware, adapters.SynchronousAdapter):
-            self._is_orig = False
             self._current = hardware
-            self._built_own_adapter = False
+        elif isinstance(hardware, ThreadManager):
+            self._current = adapters.SynchronousAdapter(hardware.managed_obj)
         else:
-            self._is_orig = False
             self._current = adapters.SynchronousAdapter(hardware)
-            self._built_own_adapter = True
 
     @property
     def hardware(self):
         return self._current
 
     def set_hw(self, hardware):
-        if self._current and (self._is_orig or self._built_own_adapter):
-            self._current.join()
         if isinstance(hardware, adapters.SynchronousAdapter):
             self._current = hardware
-            self._built_own_adapter = False
+        elif isinstance(hardware, ThreadManager):
+            self._current = hardware.sync
         elif isinstance(hardware, HardwareAPILike):
             self._current = adapters.SynchronousAdapter(hardware)
-            self._built_own_adapter = True
         else:
             raise TypeError(
                 "hardware should be API or synch adapter but is {}"
                 .format(hardware))
-        self._is_orig = False
         return self._current
 
     def reset_hw(self):
-        if self._is_orig or self._built_own_adapter:
-            self._current.join()
-        self._current = adapters.SynchronousAdapter.build(
-            API.build_hardware_simulator)
-        self._is_orig = True
-        self._built_own_adapter = True
+        self._current = ThreadManager(API.build_hardware_simulator).sync
         return self._current
-
-    def __del__(self):
-        orig = getattr(self, '_is_orig', False)
-        cur = getattr(self, '_current', None)
-        built_own = getattr(self, '_built_own_adapter')
-        if cur and (orig or built_own):
-            cur.join()
-
-    def cleanup(self):
-        """ Call to cleanup attached hardware (if it was created locally) """
-        if self._current and (self._is_orig or self._built_own_adapter):
-            self._current.join()
 
 
 def clamp_value(
