@@ -32,7 +32,7 @@ from opentrons import config, types
 from opentrons.server import init
 from opentrons.deck_calibration import endpoints
 from opentrons import hardware_control as hc
-from opentrons.hardware_control import adapters, API
+from opentrons.hardware_control import API, ThreadManager
 from opentrons.protocol_api import ProtocolContext
 from opentrons.types import Mount
 from opentrons import (robot as rb,
@@ -334,27 +334,28 @@ def virtual_smoothie_env(monkeypatch):
 @pytest.mark.skipif(aionotify is None,
                     reason="requires inotify (linux only)")
 @pytest.fixture
-def hardware(request, loop, virtual_smoothie_env):
-    hw_manager = adapters.SingletonAdapter(loop)
+async def hardware(request, loop, virtual_smoothie_env):
+    hw_sim = ThreadManager(API.build_hardware_simulator)
     old_config = config.robot_configs.load()
     try:
-        yield hw_manager
+        yield hw_sim
     finally:
-        asyncio.ensure_future(hw_manager.reset())
-        hw_manager.set_config(old_config)
+        hw_sim.set_config(old_config)
+        hw_sim.clean_up()
 
 
 @pytest.mark.skipif(aionotify is None,
                     reason="requires inotify (linux only)")
 @pytest.fixture
 def sync_hardware(request, loop, virtual_smoothie_env):
-    hardware = adapters.SynchronousAdapter.build(
-        API.build_hardware_controller)
+    thread_manager = ThreadManager(API.build_hardware_controller)
+    hardware = thread_manager.sync
     try:
         yield hardware
     finally:
         hardware.reset()
         hardware.set_config(config.robot_configs.load())
+        thread_manager.clean_up()
 
 
 @pytest.fixture
@@ -506,8 +507,8 @@ def cntrlr_mock_connect(monkeypatch):
 
 
 @pytest.fixture
-def hardware_api(loop):
-    hw_api = API.build_hardware_simulator(loop=loop)
+async def hardware_api(loop):
+    hw_api = await API.build_hardware_simulator(loop=loop)
     return hw_api
 
 
