@@ -3,14 +3,16 @@ from unittest import mock
 from opentrons.hardware_control import modules
 
 
-async def test_sim_initialization():
-    therm = await modules.build('', 'thermocycler', True, lambda x: None)
+async def test_sim_initialization(loop):
+    therm = await modules.build('', 'thermocycler', True,
+                                lambda x: None, loop=loop)
 
     assert isinstance(therm, modules.AbstractModule)
 
 
-async def test_lid():
-    therm = await modules.build('', 'thermocycler', True, lambda x: None)
+async def test_lid(loop):
+    therm = await modules.build('', 'thermocycler', True,
+                                lambda x: None, loop=loop)
 
     assert therm.lid_status == 'open'
 
@@ -27,8 +29,9 @@ async def test_lid():
     assert therm.lid_status == 'open'
 
 
-async def test_sim_state():
-    therm = await modules.build('', 'thermocycler', True, lambda x: None)
+async def test_sim_state(loop):
+    therm = await modules.build('', 'thermocycler', True,
+                                lambda x: None, loop=loop)
 
     assert therm.temperature is None
     assert therm.target is None
@@ -42,8 +45,9 @@ async def test_sim_state():
     assert status['version'] == 'dummyVersionTC'
 
 
-async def test_sim_update():
-    therm = await modules.build('', 'thermocycler', True, lambda x: None)
+async def test_sim_update(loop):
+    therm = await modules.build('', 'thermocycler', True,
+                                lambda x: None, loop=loop)
 
     await therm.set_temperature(temperature=10,
                                 hold_time_seconds=None,
@@ -82,8 +86,9 @@ async def test_sim_update():
     assert therm.lid_target is None
 
 
-async def test_set_temperature(monkeypatch):
-    hw_tc = await modules.build('', 'thermocycler', True, lambda x: None)
+async def test_set_temperature(monkeypatch, loop):
+    hw_tc = await modules.build('', 'thermocycler', True,
+                                lambda x: None, loop=loop)
 
     def async_return(result):
         f = asyncio.Future()
