@@ -60,8 +60,7 @@ def test_load_instrument(loop):
         assert ctx.loaded_instruments[Mount.RIGHT.name.lower()] == loaded
 
 
-async def test_motion(loop):
-    hardware = API.build_hardware_simulator(loop=loop)
+async def test_motion(loop, hardware):
     ctx = papi.ProtocolContext(loop)
     ctx.connect(hardware)
     ctx.home()
@@ -76,8 +75,7 @@ async def test_motion(loop):
     assert await hardware.current_position(instr._mount) == old_pos
 
 
-async def test_max_speeds(loop, monkeypatch):
-    hardware = API.build_hardware_simulator(loop=loop)
+async def test_max_speeds(loop, monkeypatch, hardware):
     ctx = papi.ProtocolContext(loop)
     ctx.connect(hardware)
     ctx.home()
@@ -104,8 +102,7 @@ async def test_max_speeds(loop, monkeypatch):
         for args, kwargs in mock_move.call_args_list)
 
 
-def test_location_cache(loop, monkeypatch, get_labware_def):
-    hardware = API.build_hardware_simulator(loop=loop)
+async def test_location_cache(loop, monkeypatch, get_labware_def, hardware):
     ctx = papi.ProtocolContext(loop)
     ctx.connect(hardware)
     right = ctx.load_instrument('p10_single', Mount.RIGHT)
@@ -140,8 +137,7 @@ def test_location_cache(loop, monkeypatch, get_labware_def):
     assert test_args[0].labware == lw.wells()[0]
 
 
-def test_move_uses_arc(loop, monkeypatch, get_labware_def):
-    hardware = API.build_hardware_simulator(loop=loop)
+async def test_move_uses_arc(loop, monkeypatch, get_labware_def, hardware):
     ctx = papi.ProtocolContext(loop)
     ctx.connect(hardware)
     ctx.home()
@@ -151,10 +147,10 @@ def test_move_uses_arc(loop, monkeypatch, get_labware_def):
 
     targets = []
 
-    async def fake_move(mount, target_pos, **kwargs):
+    async def fake_move(self, mount, target_pos, **kwargs):
         nonlocal targets
         targets.append((mount, target_pos, kwargs))
-    monkeypatch.setattr(hardware, 'move_to', fake_move)
+    monkeypatch.setattr(API, 'move_to', fake_move)
 
     right.move_to(lw.wells()[0].top())
     assert len(targets) == 3
@@ -355,9 +351,8 @@ def test_aspirate(loop, get_labware_def, monkeypatch):
 
     fake_hw_aspirate = mock.Mock()
     fake_move = mock.Mock()
-    monkeypatch.setattr(ctx._hw_manager.hardware._api,
-                        'aspirate', fake_hw_aspirate)
-    monkeypatch.setattr(ctx._hw_manager.hardware._api, 'move_to', fake_move)
+    monkeypatch.setattr(API, 'aspirate', fake_hw_aspirate)
+    monkeypatch.setattr(API, 'move_to', fake_move)
 
     instr.aspirate(2.0, lw.wells()[0].bottom())
     assert 'aspirating' in ','.join([cmd.lower() for cmd in ctx.commands()])
@@ -384,7 +379,7 @@ def test_aspirate(loop, get_labware_def, monkeypatch):
             max_speeds={})
     assert len(fake_move.call_args_list) == 2
     fake_move.reset_mock()
-    ctx._hw_manager.hardware._api\
+    ctx._hw_manager.hardware._obj_to_adapt\
                             ._attached_instruments[Mount.RIGHT]\
                             ._current_volume = 1
 
@@ -400,19 +395,18 @@ def test_dispense(loop, get_labware_def, monkeypatch):
 
     disp_called_with = None
 
-    async def fake_hw_dispense(mount, volume=None, rate=1.0):
+    async def fake_hw_dispense(self, mount, volume=None, rate=1.0):
         nonlocal disp_called_with
         disp_called_with = (mount, volume, rate)
 
     move_called_with = None
 
-    def fake_move(mount, loc, **kwargs):
+    def fake_move(self, mount, loc, **kwargs):
         nonlocal move_called_with
         move_called_with = (mount, loc, kwargs)
 
-    monkeypatch.setattr(ctx._hw_manager.hardware._api,
-                        'dispense', fake_hw_dispense)
-    monkeypatch.setattr(ctx._hw_manager.hardware._api, 'move_to', fake_move)
+    monkeypatch.setattr(API, 'dispense', fake_hw_dispense)
+    monkeypatch.setattr(API, 'move_to', fake_move)
 
     instr.dispense(2.0, lw.wells()[0].bottom())
     assert 'dispensing' in ','.join([cmd.lower() for cmd in ctx.commands()])
@@ -542,13 +536,13 @@ def test_touch_tip_default_args(loop, monkeypatch):
     instr.pick_up_tip()
     total_hw_moves = []
 
-    async def fake_hw_move(mount, abs_position, speed=None,
+    async def fake_hw_move(self, mount, abs_position, speed=None,
                            critical_point=None, max_speeds=None):
         nonlocal total_hw_moves
         total_hw_moves.append((abs_position, speed))
 
     instr.aspirate(10, lw.wells()[0])
-    monkeypatch.setattr(ctx._hw_manager.hardware._api, 'move_to', fake_hw_move)
+    monkeypatch.setattr(API, 'move_to', fake_hw_move)
     instr.touch_tip()
     z_offset = Point(0, 0, 1)   # default z offset of 1mm
     speed = 60                  # default speed
@@ -570,7 +564,7 @@ def test_touch_tip_disabled(loop, monkeypatch, get_labware_fixture):
                                 tip_racks=[tiprack])
     instr.pick_up_tip()
     move_mock = mock.Mock()
-    monkeypatch.setattr(ctx._hw_manager.hardware._api, 'move_to', move_mock)
+    monkeypatch.setattr(API, 'move_to', move_mock)
     instr.touch_tip(trough_lw['A1'])
     move_mock.assert_not_called()
 
