@@ -5,6 +5,7 @@ import pytest
 
 from opentrons import types
 from opentrons.legacy_api import modules as legacy_modules
+from opentrons.hardware_control import API
 
 from opentrons.drivers.smoothie_drivers.driver_3_0 import SmoothieDriver_3_0_0
 from opentrons.config import pipette_config
@@ -82,8 +83,9 @@ async def test_get_modules(
     hw = async_server['com.opentrons.hardware']
     magdeck = await hw._backend.build_module(port='/dev/ot_module_magdeck1',
                                              model='magdeck',
-                                             interrupt_callback=lambda x: None)
-    monkeypatch.setattr(hw, 'attached_modules', [magdeck])
+                                             interrupt_callback=lambda x: None,
+                                             loop=loop)
+    monkeypatch.setattr(API, 'attached_modules', [magdeck])
     keys = sorted(['name', 'port', 'serial', 'model', 'fwVersion',
                    'displayName', 'status', 'data', 'hasAvailableUpdate'])
     resp = await async_client.get('/modules')
@@ -94,8 +96,9 @@ async def test_get_modules(
     assert sorted(body['modules'][0].keys()) == keys
     assert 'engaged' in body['modules'][0]['data']
     tempdeck = await hw._backend.build_module('/dev/ot_module_tempdeck1',
-                                              'tempdeck', lambda x: None)
-    monkeypatch.setattr(hw, 'attached_modules', [tempdeck])
+                                              'tempdeck', lambda x: None,
+                                              loop=loop)
+    monkeypatch.setattr(API, 'attached_modules', [tempdeck])
     for model in ('temp_deck_v1', 'temp_deck_v1.1', 'temp_deck_v2'):
         tempdeck._device_info['model'] = model
         resp = await async_client.get('/modules')
@@ -129,8 +132,9 @@ async def test_execute_module_command(
 
     magdeck = await hw._backend.build_module(port='/dev/ot_module_magdeck1',
                                              model='magdeck',
-                                             interrupt_callback=lambda x: None)
-    monkeypatch.setattr(hw, 'attached_modules', [magdeck])
+                                             interrupt_callback=lambda x: None,
+                                             loop=loop)
+    monkeypatch.setattr(API, 'attached_modules', [magdeck])
 
     resp = await async_client.post('/modules/dummySerialMD',
                                    json={'command_type': 'deactivate'})
