@@ -3,7 +3,10 @@
 import * as Constants from './constants'
 import * as Types from './types'
 
-import type { RobotApiRequestMeta } from '../robot-api/types'
+import type {
+  RobotApiRequestMeta,
+  RobotApiErrorResponse,
+} from '../robot-api/types'
 
 export const fetchStatus = (robotName: string): Types.FetchStatusAction => ({
   type: Constants.FETCH_STATUS,
@@ -24,7 +27,7 @@ export const fetchStatusSuccess = (
 
 export const fetchStatusFailure = (
   robotName: string,
-  error: { ... },
+  error: RobotApiErrorResponse,
   meta: RobotApiRequestMeta
 ): Types.FetchStatusFailureAction => ({
   type: Constants.FETCH_STATUS_FAILURE,
@@ -52,7 +55,7 @@ export const fetchWifiListSuccess = (
 
 export const fetchWifiListFailure = (
   robotName: string,
-  error: { ... },
+  error: RobotApiErrorResponse,
   meta: RobotApiRequestMeta
 ): Types.FetchWifiListFailureAction => ({
   type: Constants.FETCH_WIFI_LIST_FAILURE,
@@ -81,10 +84,95 @@ export const postWifiConfigureSuccess = (
 
 export const postWifiConfigureFailure = (
   robotName: string,
-  error: { ... },
+  error: RobotApiErrorResponse,
   meta: RobotApiRequestMeta
 ): Types.PostWifiConfigureFailureAction => ({
   type: Constants.POST_WIFI_CONFIGURE_FAILURE,
+  payload: { robotName, error },
+  meta,
+})
+
+export const fetchWifiKeys = (
+  robotName: string
+): Types.FetchWifiKeysAction => ({
+  type: Constants.FETCH_WIFI_KEYS,
+  payload: { robotName },
+  meta: {},
+})
+
+export const fetchWifiKeysSuccess = (
+  robotName: string,
+  wifiKeys: Array<Types.WifiKey>,
+  meta: RobotApiRequestMeta
+): Types.FetchWifiKeysSuccessAction => ({
+  type: Constants.FETCH_WIFI_KEYS_SUCCESS,
+  payload: { robotName, wifiKeys },
+  meta,
+})
+
+export const fetchWifiKeysFailure = (
+  robotName: string,
+  error: RobotApiErrorResponse,
+  meta: RobotApiRequestMeta
+): Types.FetchWifiKeysFailureAction => ({
+  type: Constants.FETCH_WIFI_KEYS_FAILURE,
+  payload: { robotName, error },
+  meta,
+})
+
+export const postWifiKeys = (
+  robotName: string,
+  keyFile: File
+): Types.PostWifiKeysAction => ({
+  type: Constants.POST_WIFI_KEYS,
+  payload: { robotName, keyFile },
+  meta: {},
+})
+
+export const postWifiKeysSuccess = (
+  robotName: string,
+  wifiKey: Types.WifiKey,
+  meta: RobotApiRequestMeta
+): Types.PostWifiKeysSuccessAction => ({
+  type: Constants.POST_WIFI_KEYS_SUCCESS,
+  payload: { robotName, wifiKey },
+  meta,
+})
+
+export const postWifiKeysFailure = (
+  robotName: string,
+  error: RobotApiErrorResponse,
+  meta: RobotApiRequestMeta
+): Types.PostWifiKeysFailureAction => ({
+  type: Constants.POST_WIFI_KEYS_FAILURE,
+  payload: { robotName, error },
+  meta,
+})
+
+export const fetchEapOptions = (
+  robotName: string
+): Types.FetchEapOptionsAction => ({
+  type: Constants.FETCH_EAP_OPTIONS,
+  payload: { robotName },
+  meta: {},
+})
+
+export const fetchEapOptionsSuccess = (
+  robotName: string,
+  eapOptions: Array<Types.EapOption>,
+  meta: RobotApiRequestMeta
+): Types.FetchEapOptionsSuccessAction => ({
+  type: Constants.FETCH_EAP_OPTIONS_SUCCESS,
+  payload: { robotName, eapOptions },
+  meta,
+})
+
+export const fetchEapOptionsFailure = (
+  robotName: string,
+  error: RobotApiErrorResponse,
+  meta: RobotApiRequestMeta
+): Types.FetchEapOptionsFailureAction => ({
+  type: Constants.FETCH_EAP_OPTIONS_FAILURE,
   payload: { robotName, error },
   meta,
 })
