@@ -1,113 +1,81 @@
 // @flow
-import { TestScheduler } from 'rxjs/testing'
-
-import { mockRobot } from '../../../robot-api/__fixtures__'
-import * as RobotApiHttp from '../../../robot-api/http'
-import * as DiscoverySelectors from '../../../discovery/selectors'
+import { setupEpicTestMocks, runEpicTest } from '../../../robot-api/__utils__'
 import * as Fixtures from '../../__fixtures__'
 import * as Actions from '../../actions'
-import * as Types from '../../types'
 import { networkingEpic } from '..'
 
-import type { Observable } from 'rxjs'
-import type {
-  RobotHost,
-  RobotApiRequestOptions,
-  RobotApiResponse,
-} from '../../../robot-api/types'
-
-jest.mock('../../../robot-api/http')
-jest.mock('../../../discovery/selectors')
-jest.mock('../../selectors')
-
-const mockState = { state: true }
-
-const mockFetchRobotApi: JestMockFn<
-  [RobotHost, RobotApiRequestOptions],
-  Observable<RobotApiResponse>
-> = RobotApiHttp.fetchRobotApi
-
-const mockGetRobotByName: JestMockFn<[any, string], mixed> =
-  DiscoverySelectors.getRobotByName
+const makeTriggerAction = robotName => Actions.fetchStatus(robotName)
 
 describe('networking statusEpic', () => {
-  let testScheduler
-
-  beforeEach(() => {
-    mockGetRobotByName.mockReturnValue(mockRobot)
-
-    testScheduler = new TestScheduler((actual, expected) => {
-      expect(actual).toEqual(expected)
-    })
-  })
-
   afterEach(() => {
     jest.resetAllMocks()
   })
 
-  const meta = { requestId: '1234' }
-  const action: Types.FetchStatusAction = {
-    ...Actions.fetchStatus(mockRobot.name),
-    meta,
-  }
+  it('calls GET /networking/status', () => {
+    const mocks = setupEpicTestMocks(
+      makeTriggerAction,
+      Fixtures.mockNetworkingStatusSuccess
+    )
 
-  test('calls GET /networking/status', () => {
-    testScheduler.run(({ hot, cold, expectObservable, flush }) => {
-      mockFetchRobotApi.mockReturnValue(
-        cold('r', { r: Fixtures.mockNetworkingStatusSuccess })
-      )
-
-      const action$ = hot('--a', { a: action })
-      const state$ = hot('a-a', { a: mockState })
+    runEpicTest(mocks, ({ hot, expectObservable, flush }) => {
+      const action$ = hot('--a', { a: mocks.action })
+      const state$ = hot('s-s', { s: mocks.state })
       const output$ = networkingEpic(action$, state$)
 
       expectObservable(output$)
       flush()
 
-      expect(mockGetRobotByName).toHaveBeenCalledWith(mockState, mockRobot.name)
-      expect(mockFetchRobotApi).toHaveBeenCalledWith(mockRobot, {
+      expect(mocks.fetchRobotApi).toHaveBeenCalledWith(mocks.robot, {
         method: 'GET',
         path: '/networking/status',
       })
     })
   })
 
-  test('maps successful response to FETCH_STATUS_SUCCESS', () => {
-    testScheduler.run(({ hot, cold, expectObservable, flush }) => {
-      mockFetchRobotApi.mockReturnValue(
-        cold('r', { r: Fixtures.mockNetworkingStatusSuccess })
-      )
+  it('maps successful response to FETCH_STATUS_SUCCESS', () => {
+    const mocks = setupEpicTestMocks(
+      makeTriggerAction,
+      Fixtures.mockNetworkingStatusSuccess
+    )
 
-      const action$ = hot('--a', { a: action })
-      const state$ = hot('a-a', { a: {} })
+    runEpicTest(mocks, ({ hot, expectObservable }) => {
+      const action$ = hot('--a', { a: mocks.action })
+      const state$ = hot('s-s', { s: mocks.state })
       const output$ = networkingEpic(action$, state$)
 
       expectObservable(output$).toBe('--a', {
         a: Actions.fetchStatusSuccess(
-          mockRobot.name,
+          mocks.robot.name,
           Fixtures.mockNetworkingStatusSuccess.body.status,
           Fixtures.mockNetworkingStatusSuccess.body.interfaces,
-          { ...meta, response: Fixtures.mockNetworkingStatusSuccessMeta }
+          {
+            ...mocks.meta,
+            response: Fixtures.mockNetworkingStatusSuccessMeta,
+          }
         ),
       })
     })
   })
 
-  test('maps failed response to FETCH_STATUS_FAILURE', () => {
-    testScheduler.run(({ hot, cold, expectObservable, flush }) => {
-      mockFetchRobotApi.mockReturnValue(
-        cold('r', { r: Fixtures.mockNetworkingStatusFailure })
-      )
+  it('maps failed response to FETCH_STATUS_FAILURE', () => {
+    const mocks = setupEpicTestMocks(
+      makeTriggerAction,
+      Fixtures.mockNetworkingStatusFailure
+    )
 
-      const action$ = hot('--a', { a: action })
-      const state$ = hot('a-a', { a: {} })
+    runEpicTest(mocks, ({ hot, expectObservable }) => {
+      const action$ = hot('--a', { a: mocks.action })
+      const state$ = hot('s-s', { s: mocks.state })
       const output$ = networkingEpic(action$, state$)
 
       expectObservable(output$).toBe('--a', {
         a: Actions.fetchStatusFailure(
-          mockRobot.name,
+          mocks.robot.name,
           Fixtures.mockNetworkingStatusFailure.body,
-          { ...meta, response: Fixtures.mockNetworkingStatusFailureMeta }
+          {
+            ...mocks.meta,
+            response: Fixtures.mockNetworkingStatusFailureMeta,
+          }
         ),
       })
     })
