@@ -1,74 +1,37 @@
 // @flow
-import { TestScheduler } from 'rxjs/testing'
+import { setupEpicTestMocks, runEpicTest } from '../../../robot-api/__utils__'
 
-import { mockRobot } from '../../../robot-api/__fixtures__'
-import * as RobotApiHttp from '../../../robot-api/http'
 import * as Discovery from '../../../discovery'
 import * as Fixtures from '../../__fixtures__'
 import * as Actions from '../../actions'
-import * as Types from '../../types'
 import { networkingEpic } from '..'
 
-import type { Observable } from 'rxjs'
-import type {
-  RobotHost,
-  RobotApiRequestOptions,
-  RobotApiResponse,
-} from '../../../robot-api/types'
-
-jest.mock('../../../robot-api/http')
-jest.mock('../../../discovery/selectors')
-jest.mock('../../selectors')
-
-const mockState = { state: true }
-
-const mockFetchRobotApi: JestMockFn<
-  [RobotHost, RobotApiRequestOptions],
-  Observable<RobotApiResponse>
-> = RobotApiHttp.fetchRobotApi
-
-const mockGetRobotByName: JestMockFn<[any, string], mixed> =
-  Discovery.getRobotByName
-
-describe('networking wifiConfigureEpic', () => {
-  let testScheduler
-
-  beforeEach(() => {
-    mockGetRobotByName.mockReturnValue(mockRobot)
-
-    testScheduler = new TestScheduler((actual, expected) => {
-      expect(actual).toEqual(expected)
-    })
+const makeTriggerAction = robotName =>
+  Actions.postWifiConfigure(robotName, {
+    ssid: 'network-name',
+    psk: 'network-password',
   })
 
+describe('networking wifiConfigureEpic', () => {
   afterEach(() => {
     jest.resetAllMocks()
   })
 
-  const meta = { requestId: '1234' }
-  const action: Types.PostWifiConfigureAction = {
-    ...Actions.postWifiConfigure(mockRobot.name, {
-      ssid: 'network-name',
-      psk: 'network-password',
-    }),
-    meta,
-  }
-
   it('calls POST /wifi/configure with options', () => {
-    testScheduler.run(({ hot, cold, expectObservable, flush }) => {
-      mockFetchRobotApi.mockReturnValue(
-        cold('r', { r: Fixtures.mockWifiConfigureSuccess })
-      )
+    const mocks = setupEpicTestMocks(
+      makeTriggerAction,
+      Fixtures.mockWifiConfigureSuccess
+    )
 
-      const action$ = hot('--a', { a: action })
-      const state$ = hot('a-a', { a: mockState })
+    runEpicTest(mocks, ({ hot, expectObservable, flush }) => {
+      const action$ = hot('--a', { a: mocks.action })
+      const state$ = hot('s-s', { s: mocks.state })
       const output$ = networkingEpic(action$, state$)
 
       expectObservable(output$)
       flush()
 
-      expect(mockGetRobotByName).toHaveBeenCalledWith(mockState, mockRobot.name)
-      expect(mockFetchRobotApi).toHaveBeenCalledWith(mockRobot, {
+      expect(mocks.fetchRobotApi).toHaveBeenCalledWith(mocks.robot, {
         method: 'POST',
         path: '/wifi/configure',
         body: { ssid: 'network-name', psk: 'network-password' },
@@ -77,55 +40,59 @@ describe('networking wifiConfigureEpic', () => {
   })
 
   it('maps successful response to FETCH_WIFI_CONFIGURE_SUCCESS', () => {
-    testScheduler.run(({ hot, cold, expectObservable, flush }) => {
-      mockFetchRobotApi.mockReturnValue(
-        cold('r', { r: Fixtures.mockWifiConfigureSuccess })
-      )
+    const mocks = setupEpicTestMocks(
+      makeTriggerAction,
+      Fixtures.mockWifiConfigureSuccess
+    )
 
-      const action$ = hot('--a', { a: action })
-      const state$ = hot('a-a', { a: {} })
+    runEpicTest(mocks, ({ hot, expectObservable }) => {
+      const action$ = hot('--a', { a: mocks.action })
+      const state$ = hot('s-s', { s: mocks.state })
       const output$ = networkingEpic(action$, state$)
 
       expectObservable(output$).toBe('--a', {
         a: Actions.postWifiConfigureSuccess(
-          mockRobot.name,
+          mocks.robot.name,
           Fixtures.mockWifiConfigureSuccess.body.ssid,
-          { ...meta, response: Fixtures.mockWifiConfigureSuccessMeta }
+          { ...mocks.meta, response: Fixtures.mockWifiConfigureSuccessMeta }
         ),
       })
     })
   })
 
   it('maps failed response to FETCH_WIFI_CONFIGURE_FAILURE', () => {
-    testScheduler.run(({ hot, cold, expectObservable, flush }) => {
-      mockFetchRobotApi.mockReturnValue(
-        cold('r', { r: Fixtures.mockWifiConfigureFailure })
-      )
+    const mocks = setupEpicTestMocks(
+      makeTriggerAction,
+      Fixtures.mockWifiConfigureFailure
+    )
 
-      const action$ = hot('--a', { a: action })
-      const state$ = hot('a-a', { a: {} })
+    runEpicTest(mocks, ({ hot, expectObservable }) => {
+      const action$ = hot('--a', { a: mocks.action })
+      const state$ = hot('s-s', { s: mocks.state })
       const output$ = networkingEpic(action$, state$)
 
       expectObservable(output$).toBe('--a', {
         a: Actions.postWifiConfigureFailure(
-          mockRobot.name,
+          mocks.robot.name,
           Fixtures.mockWifiConfigureFailure.body,
-          { ...meta, response: Fixtures.mockWifiConfigureFailureMeta }
+          { ...mocks.meta, response: Fixtures.mockWifiConfigureFailureMeta }
         ),
       })
     })
   })
 
   it('dispatches FETCH_WIFI_LIST and START_DISCOVERY on success', () => {
-    testScheduler.run(({ hot, cold, expectObservable, flush }) => {
-      const action$ = hot('--a', {
-        a: Actions.postWifiConfigureSuccess(mockRobot.name, 'network-name', {}),
-      })
-      const state$ = hot('a-a', { a: {} })
+    const mocks = setupEpicTestMocks(robotName =>
+      Actions.postWifiConfigureSuccess(robotName, 'network-name', {})
+    )
+
+    runEpicTest(mocks, ({ hot, expectObservable }) => {
+      const action$ = hot('--a', { a: mocks.action })
+      const state$ = hot('s-s', { s: mocks.state })
       const output$ = networkingEpic(action$, state$)
 
       expectObservable(output$).toBe('--(ab)', {
-        a: Actions.fetchWifiList(mockRobot.name),
+        a: Actions.fetchWifiList(mocks.robot.name),
         b: Discovery.startDiscovery(),
       })
     })
