@@ -1,6 +1,9 @@
 // @flow
 
-import type { RobotApiRequestMeta } from '../robot-api/types'
+import type {
+  RobotApiRequestMeta,
+  RobotApiErrorResponse,
+} from '../robot-api/types'
 
 import typeof {
   FETCH_STATUS,
@@ -12,6 +15,15 @@ import typeof {
   POST_WIFI_CONFIGURE,
   POST_WIFI_CONFIGURE_SUCCESS,
   POST_WIFI_CONFIGURE_FAILURE,
+  FETCH_WIFI_KEYS,
+  FETCH_WIFI_KEYS_SUCCESS,
+  FETCH_WIFI_KEYS_FAILURE,
+  POST_WIFI_KEYS,
+  POST_WIFI_KEYS_SUCCESS,
+  POST_WIFI_KEYS_FAILURE,
+  FETCH_EAP_OPTIONS,
+  FETCH_EAP_OPTIONS_SUCCESS,
+  FETCH_EAP_OPTIONS_FAILURE,
 } from './constants'
 
 import * as ApiTypes from './api-types'
@@ -58,7 +70,7 @@ export type FetchWifiListSuccessAction = {|
 
 export type FetchWifiListFailureAction = {|
   type: FETCH_WIFI_LIST_FAILURE,
-  payload: {| robotName: string, error: { ... } |},
+  payload: {| robotName: string, error: RobotApiErrorResponse |},
   meta: RobotApiRequestMeta,
 |}
 
@@ -78,7 +90,67 @@ export type PostWifiConfigureSuccessAction = {|
 
 export type PostWifiConfigureFailureAction = {|
   type: POST_WIFI_CONFIGURE_FAILURE,
-  payload: {| robotName: string, error: { ... } |},
+  payload: {| robotName: string, error: RobotApiErrorResponse |},
+  meta: RobotApiRequestMeta,
+|}
+
+// fetch wifi keys
+
+export type FetchWifiKeysAction = {|
+  type: FETCH_WIFI_KEYS,
+  payload: {| robotName: string |},
+  meta: RobotApiRequestMeta,
+|}
+
+export type FetchWifiKeysSuccessAction = {|
+  type: FETCH_WIFI_KEYS_SUCCESS,
+  payload: {| robotName: string, wifiKeys: Array<ApiTypes.WifiKey> |},
+  meta: RobotApiRequestMeta,
+|}
+
+export type FetchWifiKeysFailureAction = {|
+  type: FETCH_WIFI_KEYS_FAILURE,
+  payload: {| robotName: string, error: RobotApiErrorResponse |},
+  meta: RobotApiRequestMeta,
+|}
+
+// post wifi keys
+
+export type PostWifiKeysAction = {|
+  type: POST_WIFI_KEYS,
+  payload: {| robotName: string, keyFile: File |},
+  meta: RobotApiRequestMeta,
+|}
+
+export type PostWifiKeysSuccessAction = {|
+  type: POST_WIFI_KEYS_SUCCESS,
+  payload: {| robotName: string, wifiKey: ApiTypes.WifiKey |},
+  meta: RobotApiRequestMeta,
+|}
+
+export type PostWifiKeysFailureAction = {|
+  type: POST_WIFI_KEYS_FAILURE,
+  payload: {| robotName: string, error: RobotApiErrorResponse |},
+  meta: RobotApiRequestMeta,
+|}
+
+// fetch eap options
+
+export type FetchEapOptionsAction = {|
+  type: FETCH_EAP_OPTIONS,
+  payload: {| robotName: string |},
+  meta: RobotApiRequestMeta,
+|}
+
+export type FetchEapOptionsSuccessAction = {|
+  type: FETCH_EAP_OPTIONS_SUCCESS,
+  payload: {| robotName: string, eapOptions: Array<ApiTypes.EapOption> |},
+  meta: RobotApiRequestMeta,
+|}
+
+export type FetchEapOptionsFailureAction = {|
+  type: FETCH_EAP_OPTIONS_FAILURE,
+  payload: {| robotName: string, error: RobotApiErrorResponse |},
   meta: RobotApiRequestMeta,
 |}
 
@@ -94,13 +166,25 @@ export type NetworkingAction =
   | PostWifiConfigureAction
   | PostWifiConfigureSuccessAction
   | PostWifiConfigureFailureAction
+  | FetchWifiKeysAction
+  | FetchWifiKeysSuccessAction
+  | FetchWifiKeysFailureAction
+  | PostWifiKeysAction
+  | PostWifiKeysSuccessAction
+  | PostWifiKeysFailureAction
+  | FetchEapOptionsAction
+  | FetchEapOptionsSuccessAction
+  | FetchEapOptionsFailureAction
 
 // state types
 
 export type PerRobotNetworkingState = $Shape<{|
-  internetStatus: ApiTypes.InternetStatus,
-  interfaces: ApiTypes.InterfaceStatusMap,
-  wifiList: Array<ApiTypes.WifiNetwork>,
+  internetStatus?: ApiTypes.InternetStatus,
+  interfaces?: ApiTypes.InterfaceStatusMap,
+  wifiList?: Array<ApiTypes.WifiNetwork>,
+  wifiKeyIds?: Array<string>,
+  wifiKeysById?: $Shape<{| [id: string]: ApiTypes.WifiKey |}>,
+  eapOptions?: Array<ApiTypes.EapOption>,
 |}>
 
 export type NetworkingState = $Shape<{|
