@@ -24,7 +24,7 @@ export type ResponseFixtures<SuccessBody, FailureBody> = {|
     method: Method,
     path: string,
     status: number,
-    ok: true,
+    ok: boolean,
     host: RobotHost,
     body: SuccessBody,
   |},
@@ -32,7 +32,7 @@ export type ResponseFixtures<SuccessBody, FailureBody> = {|
     method: Method,
     path: string,
     status: number,
-    ok: false,
+    ok: boolean,
     host: RobotHost,
     body: FailureBody,
   |},
